Move GetTempFilename to tensorflow/core/lib/io/path.h
